requirements.txt
asgiref==3.6.0
Django==4.2.1
sqlparse==0.4.4
djangorestframework==3.14.0
